Fix warnings from controller side. Reason for change: 1) Handled some datatype conversion related warnings 2) Handled warnings related to makefile flags and seperated C and CPP flags Test Procedure: Checked and compared the warning before and after fix, observed around 94 warning is fixed from controller compilation. Also compared the run time logs for OneWifi, controller and agent with and without changes and did not observed any additional logs/error. Tested the changes by changing the SSID, Passphrase and radio channel through CLI observed its works fine. Risks: Medium Priority: P1

+++ src/ctrl/dm_easy_mesh_ctrl.cpp
@@ -718,11 +718,11 @@ int dm_easy_mesh_ctrl_t::analyze_set_channel(em_bus_event_t *evt, em_cmd_t *pcmd
 					// Check if the channel has changed or not
 					if (updated_oclass->m_op_class_info.channels[0] != current_oclass->m_op_class_info.channels[0]) {
 						already_added = 0;
-						band = dm_easy_mesh_t::get_freq_band_by_op_class(updated_oclass->m_op_class_info.id.op_class);
+						band = dm_easy_mesh_t::get_freq_band_by_op_class(static_cast<int>(updated_oclass->m_op_class_info.id.op_class));
 
 						// Check if the band is already added to event parameters
 						for (k = 0; k < evt->params.u.args.num_args; k++) {
-							if (atoi(evt->params.u.args.args[k]) == band) {
+							if (static_cast<unsigned int>(atoi(evt->params.u.args.args[k])) == band) {
 								already_added = 1;
 								break;
 							}
